tsinghua-auth/requirements.txt
requests
pyannotate
pyyaml
